deployment/requirements.txt
alpaca-py
aiohttp
python-dotenv
sqlalchemy
pandas
numpy